+++ src/serve.rs
@@ -2,6 +2,9 @@
 use std::sync::Arc;
 
 use tokio::net::TcpListener;
+#[cfg(feature = "tracing")]
+extern crate tracing;
+#[cfg(feature = "tracing")]
 use tracing::{event, Level};
 
 use crate::{router::Router, AGIError};
@@ -15,6 +18,7 @@ pub async fn serve(listener: TcpListener, router: Router) -> Result<(), AGIError
             .accept()
             .await
             .map_err(|_| AGIError::CannotSpawnListener)?;
+        #[cfg(feature = "tracing")]
         event!(Level::DEBUG, "Got a new incoming connection.");
         tokio::spawn(async move {
             our_router.handle(stream).await;